+++ NEWS.md
@@ -5,7 +5,12 @@
 # convdistr 0.2.0
 
 * DISTRIBUTION_factory added
-* new_BINOMIAL function added
+* New_BINOMIAL function added
+* Added the fitdiri and fitbeta functions
+* Added the dimmnames function
+* All distributions now accept a dimnames paramater
+* Adds  MULTINORMAL distribution
+